+++ Scenes/UI/StationMenu/TradeMenu/TradeAction/TradeAction.cs
@@ -1,7 +1,7 @@
 using System;
 using Godot;
 using Serilog;
-using Spacelancer.Components.Economy.Commodities;
+using Spacelancer.Economy;
 
 namespace Spacelancer.Scenes.UI.StationMenu.TradeMenu.TradeAction;
