+++ Scenes/UI/StationMenu/TradeMenu/TradeMenu.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Godot;
 using Serilog;
-using Spacelancer.Components.Economy.Commodities;
+using Spacelancer.Economy;
 using Spacelancer.Components.Storage;
 
 namespace Spacelancer.Scenes.UI.StationMenu.TradeMenu;
